+++ shopil/src/main/java/com/seagroup/seatalk/shopil/decode/Decoder.kt
@@ -1,4 +1,4 @@
-package com.seagroup.seatalk.shopil.decoder
+package com.seagroup.seatalk.shopil.decode
 
 interface Decoder {
     suspend fun decode(params: DecodeParams): DecodeResult
