+++ shopil/src/main/java/com/seagroup/seatalk/shopil/fetch/FileFetcher.kt
@@ -1,4 +1,4 @@
-package com.seagroup.seatalk.shopil.fetcher
+package com.seagroup.seatalk.shopil.fetch
 
 import com.seagroup.seatalk.shopil.request.ImageSource
 import okio.buffer
