+++ shopil/src/main/java/com/seagroup/seatalk/shopil/fetch/FileUriFetcher.kt
@@ -1,4 +1,4 @@
-package com.seagroup.seatalk.shopil.fetcher
+package com.seagroup.seatalk.shopil.fetch
 
 import androidx.core.net.toFile
 import com.seagroup.seatalk.shopil.request.ImageSource
